# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: API Gestor de Tareas
  version: 1.0.0
servers:
  - url: /api

components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # <<< TODO: Check >>>
    AuthRequest:
      type: object
      required: [ username, password ]
      properties:
        username:
          type: string
          example: esteban
        password:
          type: string
          example: admin123

    AuthResponse:
      type: object
      required: [ access_token, expires_at ]
      properties:
        access_token:
          type: string
          description: Token JWT
        expires_at:
          type: string
          format: date-time
          description: Fecha/hora de expiración (ISO-8601)

    TaskDto:
      type: object
      required: [ title, statusId ]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        statusId:
          type: integer
          format: int64

paths:
  /auth/login:
    post:
      summary: Autenticarse y recibir JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'   # ← ahora existe
      responses:
        '200':
          description: Token JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse' # ← ahora existe

  /tasks:
    get:
      security: [ { jwtAuth: [ ] } ]
      summary: Listar tareas
      responses:
        '200':
          description: Lista de tareas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDto'
    post:
      security: [ { jwtAuth: [ ] } ]
      summary: Crear tarea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskDto'
      responses:
        '201':
          description: Creada

#check language